SDK-553: Split yotiprofile.go into application_profile.go and baseprofile.go
