+++ baseprofile.go
@@ -1,0 +1,51 @@
+package yoti
+
+import (
+	"github.com/getyoti/yoti-go-sdk/v2/attribute"
+	"github.com/getyoti/yoti-go-sdk/v2/yotiprotoattr"
+)
+
+type baseProfile struct {
+	attributeSlice []*yotiprotoattr.Attribute
+}
+
+// GetAttribute retrieve an attribute by name on the Yoti profile. Will return nil if attribute is not present.
+func (p baseProfile) GetAttribute(attributeName string) *attribute.GenericAttribute {
+	for _, a := range p.attributeSlice {
+		if a.Name == attributeName {
+			return attribute.NewGeneric(a)
+		}
+	}
+	return nil
+}
+
+func (p baseProfile) GetStringAttribute(attributeName string) *attribute.StringAttribute {
+	for _, a := range p.attributeSlice {
+		if a.Name == attributeName {
+			return attribute.NewString(a)
+		}
+	}
+	return nil
+}
+
+func (p baseProfile) GetImageAttribute(attributeName string) *attribute.ImageAttribute {
+	for _, a := range p.attributeSlice {
+		if a.Name == attributeName {
+			attribute, err := attribute.NewImage(a)
+
+			if err == nil {
+				return attribute
+			}
+		}
+	}
+	return nil
+}
+
+func (p baseProfile) GetJSONAttribute(attributeName string) (*attribute.JSONAttribute, error) {
+	for _, a := range p.attributeSlice {
+		if a.Name == attributeName {
+			return attribute.NewJSON(a)
+		}
+	}
+	return nil, nil
+}
